fix for aarch64

+++ setup.sh
@@ -3,6 +3,16 @@
 set -e
 set -a
 
+DEPLOY_CLIENT=true
+
+# If x86_64 deploy Linux client app and RFID bridge
+# Else, deploy only RFID bridge, app is not currently available for other architectures
+# You can access the web client at http://openpark.com
+if [[ "$(uname -m)" != "x86_64" ]]; then
+    DEPLOY_CLIENT=false
+    echo "Open Park Project client app is not available for this architecture. Only RFID bridge will be deployed."
+fi
+
 CLIENT_SERVICE_NAME="opp-client.service"
 USER_NAME=$USER
 CLIENT_EXECUTABLE_DIR=$HOME/opp-frontend/
@@ -26,16 +36,18 @@ if [ -d "$RFID_BRIDGE_EXECUTABLE_DIR" ]; then
     rm -rf $RFID_BRIDGE_EXECUTABLE_DIR
 fi
 
-echo "Fetching latest client app..."
-# Download the latest OPP client frontend
-OPP_CLIENT_URL_BASE="https://api.github.com/repos/OpenParkProject/OPP-frontend/releases/latest"
-DOWNLOAD_URL=$(curl -s $OPP_CLIENT_URL_BASE | grep "browser_download_url" | grep "opp-linux-.*\.tar\.gz" | cut -d '"' -f 4)
+if [ $DEPLOY_CLIENT = true ]; then
+  echo "Fetching latest client app..."
+  # Download the latest OPP client frontend
+  OPP_CLIENT_URL_BASE="https://api.github.com/repos/OpenParkProject/OPP-frontend/releases/latest"
+  DOWNLOAD_URL=$(curl -s $OPP_CLIENT_URL_BASE | grep "browser_download_url" | grep "opp-linux-.*\.tar\.gz" | cut -d '"' -f 4)
 
-echo "Downloading from $DOWNLOAD_URL"
-curl -s -L -o /tmp/opp-client.tar.gz "$DOWNLOAD_URL"
-mkdir -p $CLIENT_EXECUTABLE_DIR
-tar -xzf /tmp/opp-client.tar.gz -C $CLIENT_EXECUTABLE_DIR --strip-components=1
-rm -f /tmp/opp-client.tar.gz
+  echo "Downloading from $DOWNLOAD_URL"
+  curl -s -L -o /tmp/opp-client.tar.gz "$DOWNLOAD_URL"
+  mkdir -p $CLIENT_EXECUTABLE_DIR
+  tar -xzf /tmp/opp-client.tar.gz -C $CLIENT_EXECUTABLE_DIR --strip-components=1
+  rm -f /tmp/opp-client.tar.gz
+fi
 
 echo "Installing OPP RFID Bridge..."
 mkdir -p $RFID_BRIDGE_EXECUTABLE_DIR
@@ -50,23 +62,23 @@ pip install -r $RFID_BRIDGE_EXECUTABLE_DIR/requirements.txt
 deactivate
 
 echo "Creating systemd service files..."
-sudo cp -f services/$CLIENT_SERVICE_NAME /etc/systemd/system/
-sudo cp -f services/$RFID_BRIDGE_SERVICE_NAME /etc/systemd/system/
-
 # Substitute environment variables in service files
-echo "Creating systemd service files..."
-envsubst < services/$CLIENT_SERVICE_NAME > /tmp/$CLIENT_SERVICE_NAME
 envsubst < services/$RFID_BRIDGE_SERVICE_NAME > /tmp/$RFID_BRIDGE_SERVICE_NAME
-sudo cp /tmp/$CLIENT_SERVICE_NAME /etc/systemd/system/
 sudo cp /tmp/$RFID_BRIDGE_SERVICE_NAME /etc/systemd/system/
+if [ $DEPLOY_CLIENT = true ]; then
+  envsubst < services/$CLIENT_SERVICE_NAME > /tmp/$CLIENT_SERVICE_NAME
+  sudo cp /tmp/$CLIENT_SERVICE_NAME /etc/systemd/system/
+fi
 
 # Clean up temporary files
 rm /tmp/$CLIENT_SERVICE_NAME /tmp/$RFID_BRIDGE_SERVICE_NAME
 echo "Enabling and starting services..."
 sudo systemctl daemon-reload
-sudo systemctl enable $CLIENT_SERVICE_NAME
 sudo systemctl enable $RFID_BRIDGE_SERVICE_NAME
-sudo systemctl start $CLIENT_SERVICE_NAME
 sudo systemctl start $RFID_BRIDGE_SERVICE_NAME
+if [ $DEPLOY_CLIENT = true ]; then
+  sudo systemctl enable $CLIENT_SERVICE_NAME
+  sudo systemctl start $CLIENT_SERVICE_NAME
+fi
 
 echo "Open Park Project totem configuration completed successfully!"
